requirements.txt
pandas
numpy
matplotlib
scikit-learn
tensorflow==2.16.*
pyyaml
requests
shapely
geopandas
fastapi
uvicorn
python-multipart
